% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uas2csv.R
\name{uas2csv}
\alias{uas2csv}
\title{Merge and widen files generated from ForenSeq UAS and export as a single CSV file}
\usage{
uas2csv(files = files, population = pop_file, reference = FALSE)
}
\arguments{
\item{files}{is a zipped or tar file containing all individual/group reports}

\item{population}{is a csv or excel file containing the metadata. It should have a "Sample" header.}

\item{reference}{indicates if there is a "population" input. Should be set to TRUE if metadata is uploaded. Default is FALSE.}
}
\description{
Merge and widen files generated from ForenSeq UAS and export as a single CSV file
}
\examples{
uas2csv(file = "files.zip", population = "metadata.xlsx", reference = TRUE)
}
